Function hpke_aead::AeadSeal

pub fn AeadSeal(
    aead_id: AEAD, 
    key: &Key, 
    nonce: &Nonce, 
    aad: &Seq<U8>, 
    pt: &Seq<U8>
) -> HpkeByteSeqResult

Encrypt and authenticate plaintext pt with associated data aad using symmetric key key and nonce nonce, yielding ciphertext and tag ct. This function can raise a MessageLimitReachedError upon failure.